// src/app/interfaces/MateriaPrima.interface.ts
export default interface MateriaPrima{
    flag_nueva_materia_prima?: boolean;
    codigo_materia_prima?: string;
    descripcion_materia_prima?: string;
    SUPPLIER_RESIN?: string;
    parte_g?: string;
    RUNNER_percent?: string;
    RUNNER_g?: string;
    loss_percent?: string;
    weigth?: string;
    resin_basis?: string;
    overhead_cost?: string;
    resin_cot?: string;
    total_mat?: string;
    GK_percent?: string;
    GK?:string;
    scrap_percent?: string;
    margin_scrap?: string;
    margin_2?: string;
    margen_seguridad_percent?: string;
    margen_seguridad?: string;
}